externals: Update LibCommon

+++ src/Core/Resource/Factory/CAnimationLoader.cpp
@@ -462,9 +462,9 @@ CQuaternion CAnimationLoader::DequantizeRotation(bool Sign, int16 X, int16 Y, in
     const float Multiplier = Math::skHalfPi / static_cast<float>(mRotationDivisor);
 
     CQuaternion Out;
-    Out.X = std::sinf(static_cast<float>(X) * Multiplier);
-    Out.Y = std::sinf(static_cast<float>(Y) * Multiplier);
-    Out.Z = std::sinf(static_cast<float>(Z) * Multiplier);
+    Out.X = std::sin(static_cast<float>(X) * Multiplier);
+    Out.Y = std::sin(static_cast<float>(Y) * Multiplier);
+    Out.Z = std::sin(static_cast<float>(Z) * Multiplier);
     Out.W = std::sqrt(std::fmax(1.f - ((Out.X * Out.X) + (Out.Y * Out.Y) + (Out.Z * Out.Z)), 0.f));
 
     if (Sign)
